backend/requirements.txt
absl-py==2.1.0
agno==1.1.3
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.18.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
duckdb==1.2.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
ipykernel==6.29.5
ipython==8.35.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-deepseek==0.1.2
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langsmith==0.1.147
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.5.1
mistune==3.1.3
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.26.4
openai==1.63.1
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pgvector==0.3.6
pillow==11.1.0
platformdirs==4.3.7
portalocker==3.1.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.3
pypdf==5.3.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-json-logger==3.3.0
python-multipart==0.0.20
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.4.0
ragas==0.2.14
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rouge-score==0.1.2
rpds-py==0.24.0
sacrebleu==2.5.1
seaborn==0.13.2
Send2Trash==1.8.3
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
